Revert "Merge pull request #20 from Gabo8912/dilan"

+++ MtdrSpring/backend/src/main/java/com/springboot/MyTodoList/service/TaskAssignmentService.java
@@ -8,34 +8,31 @@ import com.springboot.MyTodoList.repository.ToDoItemRepository;
 import com.springboot.MyTodoList.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import java.util.List;
 import java.util.Optional;
-
 @Service
 public class TaskAssignmentService {
 
     @Autowired
     private TaskAssignmentRepository taskAssignmentRepository;
-
+    
     @Autowired
     private ToDoItemRepository toDoItemRepository;
-
+    
     @Autowired
     private UserRepository userRepository;
 
     public TaskAssignment createAssignment(Long taskId, Long userId) {
-        // Convert Long to Integer for taskId, but keep userId as Long
+        // Convert Long to Integer for repository calls
         Optional<ToDoItem> task = toDoItemRepository.findById(taskId.intValue());
-        Optional<User> user = userRepository.findById(userId); // ✅ esto ya está correcto
-
+        Optional<User> user = userRepository.findById(userId.intValue());
+        
         if (task.isPresent() && user.isPresent()) {
             TaskAssignment newAssignment = new TaskAssignment();
             newAssignment.setTask(task.get());
             newAssignment.setUser(user.get());
             return taskAssignmentRepository.save(newAssignment);
         }
-
         throw new IllegalArgumentException("Task or User not found with provided IDs");
     }
